+++ backpack.tf extended sorting.meta.js
@@ -1,10 +1,10 @@
 // ==UserScript==
 // @name         backpack.tf - Miscellaneous Extensions
 // @description  Adds more options for sorting items in backpacks (currently Sorting for paints, spells, levels, scm price, classified listings) and other stuff which I would have liked (including highlighting spells, autocompleting spell names, autocompleting particle names or sorting unusuals by class)
-// @version      0.1.26
+// @version      0.1.27
 // @author       Netroscript
 // @namespace    https://github.com/NetroScript
-// @include      /^https?:\/\/backpack\.tf\/.*
+// @include      /^https?:\/\/(.*\.)?backpack\.tf\/.*
 // @downloadURL  https://github.com/NetroScript/backpack.tf-miscellaneous-extensions/raw/master/backpack.tf%20extended%20sorting.user.js
 // @updateURL    https://github.com/NetroScript/backpack.tf-miscellaneous-extensions/raw/master/backpack.tf%20extended%20sorting.meta.js
 // @grant        none